transform to ts AND remove old state
